+++ src/wine-host/bridges/vst2.cpp
@@ -94,10 +94,15 @@ static const std::unordered_set<int> safe_mutually_recursive_requests{
  * NOTE: `effSetChunk` and `effGetChunk` should be callable from any thread, but
  *       Algonaut Atlas doesn't restore chunk data unless `effSetChunk` is run
  *       from the GUI thread
+ * NOTE: `effSetSampleRate` and `effSetBlockSize` really shouldn't be here, but
+ *       New Sonic Arts' Vice plugin spawns a new thread and calls drawing code
+ *       while changing sample rate and block size. We'll need to see if doing
+ *       this on the main thread introduces any regressions.
  */
 static const std::unordered_set<int> unsafe_requests{
-    effOpen,     effClose,   effEditGetRect,  effEditOpen, effEditClose,
-    effEditIdle, effEditTop, effMainsChanged, effGetChunk, effSetChunk};
+    effOpen,      effClose,    effEditGetRect,   effEditOpen,
+    effEditClose, effEditIdle, effEditTop,       effMainsChanged,
+    effGetChunk,  effSetChunk, effSetSampleRate, effSetBlockSize};
 
 /**
  * These opcodes from `unsafe_requests` should be run under realtime scheduling